+++ packages/framework/src/Concerns/HasTableOfContents.php
@@ -2,10 +2,12 @@
 
 namespace Hyde\Framework\Concerns;
 
-use Hyde\Framework\Actions\GeneratesTableOfContents;
+use Hyde\Framework\Actions\GeneratesSidebarTableOfContents;
 
 /**
  * Generate Table of Contents as HTML from a Markdown document body.
+ * 
+ * Intended to be used for documentation pages.
  *
  * @see \Hyde\Framework\Testing\Unit\HasTableOfContentsTest
  */
@@ -13,6 +15,6 @@ trait HasTableOfContents
 {
     public function getTableOfContents(): string
     {
-        return (new GeneratesTableOfContents($this->body))->execute();
+        return (new GeneratesSidebarTableOfContents($this->body))->execute();
     }
 }